+++ CHANGELOG.md
@@ -5,7 +5,7 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 -
 -
 -
--
+- Fix graph point overlaps
 -
 -
 -