+++ CHANGELOG.md
@@ -5,10 +5,13 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 -
 -
 -
+- Fix graph point overlaps
 -
 -
 -
 - Adds avg_net_price to scorecard api updates
+- Added animation to button interactions
+-
 
 ## 2.0.1
 - Updated content to be dynamic based on remaining cost