+++ CHANGELOG.md
@@ -11,7 +11,9 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 -
 - Adds avg_net_price to scorecard api updates
 - Added animation to button interactions
--
+- Switched to `merge=union` for CHANGELOG.md in .gitattributes
+- 
+- 
 
 ## 2.0.1
 - Updated content to be dynamic based on remaining cost