+++ CHANGELOG.md
@@ -5,10 +5,11 @@ We follow the [Semantic Versioning 2.0.0](http://semver.org/) format.
 -
 -
 -
+-
+
+## 2.0.2
+- Add active class to selected big question button
 - Fix graph point overlaps
--
--
--
 - Adds avg_net_price to scorecard api updates
 - Added animation to button interactions
 - Switched to `merge=union` for CHANGELOG.md in .gitattributes